requirements.txt
kafka-python
tweepy
transformers
torch
pandas
